feat: 유저 등록

+++ src/main/java/com/ebcho/engdabot/repository/TelegramUserRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ebcho.engdabot.entity.TelegramUser;
 
-public interface TelegramUserRepository extends CrudRepository<TelegramUser, Integer> {
+public interface TelegramUserRepository extends CrudRepository<TelegramUser, Long> {
 }
