+++ src/main/java/com/ebcho/engdabot/telegram/User.java
@@ -3,7 +3,7 @@ package com.ebcho.engdabot.telegram;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record User(
-	@JsonProperty("id") Integer id,
+	@JsonProperty("id") Long id,
 	@JsonProperty("is_bot") Boolean isBot,
 	@JsonProperty("first_name") String firstName,
 	@JsonProperty("username") String username,
